Change how replaygain tags are handled.
Continue to copy the REPLAYGAIN_ALBUM_GAIN to the OpusHead gain value.
Screw the R128_TRACK_GAIN tag, nothing supports it.
Instead set REPLAYGAIN_ALBUM_GAIN to 0 so players know not to apply
fallback gain, and adjust REPLAYGAIN_TRACK_GAIN accordingly.
Use the replaygain reference loudness of 84DBFS instead of
the EBU.R128 reference of 89DBFS, which is too quiet.

Also tweak the Makefile a bit for my personal setup.